src/requirements.txt
Django==2.2.13
Pygments==2.4.2
Markdown==3.1.1
-e git+https://github.com/NTIA/SigMF.git@multi-recording-archive#egg=SigMF
coreapi==2.3.3
django-debug-toolbar==1.11
django-extensions==2.2.6
django-filter==2.2.0
djangorestframework==3.9.4
docker-compose==1.24.0
drf-yasg==1.15.0
environs==4.2.0
gunicorn==19.9.0
jsonfield==2.0.2
numpy>=1.16.4
oauthlib~=3.1.0
psycopg2-binary==2.8.5
pyjwt~=1.7.1
python-dateutil==2.8.1
raven==6.10.0
requests-futures==0.9.9
requests-mock==1.6.0
requests-oauthlib~=1.3.0
ruamel.yaml==0.15.96
scipy==1.3.1
six==1.12.0
typing==3.6.6
